+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(LandR.CS)
+
+test_check("LandR.CS")